requirements.txt
setuptools==67.6.1
networkx==3.1
sortedcontainers~=2.4.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
google-genai>=0.2.0
requests>=2.31.0